+++ @xen-orchestra/web/typed-router.d.ts
@@ -6,13 +6,7 @@
 // Make sure to add this file to your tsconfig.json file as an "includes" or "files" entry.
 
 declare module 'vue-router/auto-routes' {
-  import type {
-    RouteRecordInfo,
-    ParamValue,
-    ParamValueOneOrMore,
-    ParamValueZeroOrMore,
-    ParamValueZeroOrOne,
-  } from 'vue-router'
+  import type { ParamValue, RouteRecordInfo, } from 'vue-router'
 
   /**
    * Route name map generated by unplugin-vue-router
