+++ packages/xo-server/src/db-cli.mjs
@@ -1,28 +1,10 @@
 #!/usr/bin/env node
 
 import { start as startRepl } from 'repl'
-import { Worker } from 'worker_threads'
 import appConf from 'app-conf'
+import { startSpinner } from './_cli-utils.mjs'
 
 import RedisCollection from './collection/redis.mjs'
-
-function startSpinner() {
-  if (!process.stdout.isTTY) return () => {}
-
-  const worker = new Worker(
-    `const { parentPort } = require('worker_threads')
-    const { writeSync } = require('fs')
-    const frames = [' | ', ' / ', ' - ', ' \\\\ ']
-    let i = 0
-    writeSync(1, frames[0])
-    const id = setInterval(() => writeSync(1, '\\r' + frames[i++ % frames.length]), 100)
-    parentPort.once('message', () => { clearInterval(id); writeSync(1, '\\r\\x1b[K') })`,
-    { eval: true }
-  )
-  worker.on('error', () => {})
-
-  return () => worker.postMessage('stop')
-}
 
 function assert(test, message) {
   if (!test) {
