+++ packages/xo-server/src/recover-account-cli.mjs
@@ -1,29 +1,11 @@
 #!/usr/bin/env node
 
-import { Worker } from 'worker_threads'
 import appConf from 'app-conf'
 import execPromise from 'exec-promise'
 import pw from 'pw'
+import { startSpinner } from './_cli-utils.mjs'
 
 import { generateToken } from './utils.mjs'
-
-function startSpinner() {
-  if (!process.stdout.isTTY) return () => {}
-
-  const worker = new Worker(
-    `const { parentPort } = require('worker_threads')
-    const { writeSync } = require('fs')
-    const frames = [' | ', ' / ', ' - ', ' \\\\ ']
-    let i = 0
-    writeSync(1, frames[0])
-    const id = setInterval(() => writeSync(1, '\\r' + frames[i++ % frames.length]), 100)
-    parentPort.once('message', () => { clearInterval(id); writeSync(1, '\\r\\x1b[K') })`,
-    { eval: true }
-  )
-  worker.on('error', () => {})
-
-  return () => worker.postMessage('stop')
-}
 
 execPromise(async args => {
   if (args.length === 0 || args.length > 2 || args[0] === '-h' || args[0] === '--help') {
